Change counting technique

+++ src/statement_pruner.cpp
@@ -66,7 +66,7 @@ remove_statements(const IR::P4Program *temp,
 
 const IR::P4Program *prune_statements(const IR::P4Program *program,
                                       P4PRUNER::PrunerConfig pruner_conf,
-                                      double prog_size) {
+                                      uint64_t prog_size) {
     int same_before_pruning = 0;
     int result;
     int max_statements = prog_size / SIZE_BANK_RATIO;
